Partial matching of Kitchen instance names
By parsing `.kitchen.yml` for a list of all available instances and
searching the list for the provided instance name using partial
matching.

Partial matching allows the user to specify only a portion of the
instance name. For example, `chef-runner -K ubuntu` will provision the
instance "default-ubuntu-1404" if that's the first instance with
"ubuntu" in its name.

Fixes #23.